CRUD v1

+++ src/main/java/com/jakurych/blockplan/blockplan/service/UserService.java
@@ -1,11 +1,12 @@
 package com.jakurych.blockplan.blockplan.service;
-import com.jakurych.blockplan.blockplan.controller.UserController;
+
 import com.jakurych.blockplan.blockplan.model.domain.User;
 import com.jakurych.blockplan.blockplan.model.repository.UserRepository;
-import lombok.Builder;
+import lombok.Builder; // Import the @Builder annotation
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 @Service
@@ -15,25 +16,34 @@ public class UserService {
     private final UserRepository userRepository;
 
     public User createOrUpdate(User user) {
-        userRepository.save(user);
-        return null; //TODO: one/two line(s)
+        return userRepository.save(user);
     }
 
     public User delete(Integer id) {
         User user = findById(id);
-        userRepository.delete(user);
-        return user; //TODO: one/two line(s)
+        if (user != null) {
+            userRepository.delete(user);
+        }
+        return user;
     }
 
     public User findById(int id) {
-
         return userRepository.findById(id).orElse(null);
     }
-    public User findByUsername(String username){
+
+    public User findByUsername(String username) {
         return userRepository.findByUsername(username);
     }
 
     public List<User> findAll() {
         return userRepository.findAll();
+    }
+
+    public User findByEmail(String email) {
+        return userRepository.findByEmail(email);
+    }
+
+    public List<User> findByRegistrationDateBetween(LocalDateTime startDateTime, LocalDateTime endDateTime) {
+        return userRepository.findByRegistrationDateBetween(startDateTime, endDateTime);
     }
 }
